Task002

+++ src/main/java/dao/DAOUsuarioRepository.java
@@ -17,75 +17,89 @@ public class DAOUsuarioRepository {
 		connection = SingleConnectionBanco.getConnection();
 	}
 
-	public ModelLogin gravarUsuario(ModelLogin objeto) throws Exception {
+	public ModelLogin gravarUsuario(ModelLogin objeto, Long userLogado) throws Exception {
 
 		if(objeto.isNovo()) {// Gravar um novo
-			String sql = "INSERT INTO model_login(login, senha, nome, email) VALUES (?,?,?,?);";
+			String sql = "INSERT INTO model_login(login, senha, nome, email, usuario_id, perfil, sexo) VALUES (?,?,?,?,?,?,?);";
 			PreparedStatement preparedSql = connection.prepareStatement(sql);
 			
 			preparedSql.setString(1, objeto.getLogin());
 			preparedSql.setString(2, objeto.getSenha());
 			preparedSql.setString(3, objeto.getNome());
 			preparedSql.setString(4, objeto.getEmail());
+			preparedSql.setLong(5, userLogado);
+			preparedSql.setString(6, objeto.getPerfil());
+			preparedSql.setString(7, objeto.getSexo());
 			preparedSql.execute();
 	
 			connection.commit();// Commita no banco.
+			
+			if(objeto.getFotouser() != null && !objeto.getFotouser().isEmpty()) {
+			
+				sql = "UPDATE model_login set fotouser=?, extensaofotouser=? WHERE login =?";
+				
+				preparedSql = connection.prepareStatement(sql);
+				
+				preparedSql.setString(1, objeto.getFotouser());
+				preparedSql.setString(2, objeto.getExtensaofotouser());
+				preparedSql.setString(3, objeto.getLogin());
+				
+				preparedSql.execute();
+				
+				connection.commit();
+			}
+			
 		}else {// Update 
-			String sql = "UPDATE model_login SET login=?, senha=?, nome=?, email=? WHERE id = "+objeto.getId()+";";
+			String sql = "UPDATE model_login SET login=?, senha=?, nome=?, email=?, perfil=?, sexo=? WHERE id = "+objeto.getId()+";";
 			PreparedStatement prepareSql = connection.prepareStatement(sql);
 			
 			prepareSql.setString(1, objeto.getLogin());
 			prepareSql.setString(2, objeto.getSenha());
 			prepareSql.setString(3, objeto.getNome());
 			prepareSql.setString(4, objeto.getEmail());
+			prepareSql.setString(5, objeto.getPerfil());
+			prepareSql.setString(6, objeto.getSexo());
 			prepareSql.executeUpdate();
 			
 			connection.commit();
+			
+			if(objeto.getFotouser() != null && !objeto.getFotouser().isEmpty()) {
+				
+				sql = "update model_login set fotouser=?, extensaofotouser=? WHERE id =?";
+				
+				prepareSql = connection.prepareStatement(sql);
+				
+				prepareSql.setString(1, objeto.getFotouser());
+				prepareSql.setString(2, objeto.getExtensaofotouser());
+				prepareSql.setLong(3, objeto.getId());
+				
+				prepareSql.execute();
+				
+				connection.commit();
+			}
 		}
 		
-		return this.consultaUsuarioList(objeto.getLogin());
+		return this.consultarUsuario(objeto.getLogin());
 	}
 
-	
-	public List<ModelLogin> consultarUserAjax(String nome) throws Exception {
-		
-		List<ModelLogin> retorno = new ArrayList<>();
-		
-		String sql = "select * from model_login where upper(nome) like upper(?) ";
-		PreparedStatement statement = connection.prepareStatement(sql);
-		statement.setString(1, "%" + nome + "%");
-		
-		ResultSet resultado = statement.executeQuery();
-		while(resultado.next()) {// Percorrer a linha de resultado do SQL.
-			ModelLogin modelLogin = new ModelLogin();
-			modelLogin.setId(resultado.getLong("id"));
-			modelLogin.setNome(resultado.getString("nome"));
-			modelLogin.setEmail(resultado.getString("email"));
-			modelLogin.setLogin(resultado.getString("login"));
-			
-			retorno.add(modelLogin);
-		}
-		
-		return retorno;
-	}
-	
-	
-	public List<ModelLogin> consultaUsuarioList() throws Exception {
+	// Consulta todos
+	public List<ModelLogin> consultaUsuarioList(Long userLogado) throws Exception {
 		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
 
-		String sql = "select * from model_login";
+		String sql = "select * from model_login where useradmin is false and usuario_id = " + userLogado;
 		PreparedStatement statement = connection.prepareStatement(sql);
 
 		ResultSet resultado = statement.executeQuery();
 
 		while (resultado.next()) { /* percorrer as linhas de resultado do SQL */
-
 			ModelLogin modelLogin = new ModelLogin();
 
 			modelLogin.setId(resultado.getLong("id"));
 			modelLogin.setNome(resultado.getString("nome"));
 			modelLogin.setEmail(resultado.getString("email"));
 			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setPerfil(resultado.getString("perfil"));
+			modelLogin.setSexo(resultado.getString("sexo"));
 
 			retorno.add(modelLogin);
 		}
@@ -93,12 +107,66 @@ public class DAOUsuarioRepository {
 		return retorno;
 	}
 	
+	// Consulta a lista
+	public List<ModelLogin> consultaUsuarioList(String nome, Long userLogado) throws Exception {
+		List<ModelLogin> retorno = new ArrayList<ModelLogin>();
+			
+		ModelLogin modelLogin = new ModelLogin();
+
+		String sql = "select * from model_login where upper(nome) like upper(?) and useradmin is false and usuario_id = ?";
+
+		PreparedStatement statement = connection.prepareStatement(sql);
+		statement.setString(1, "%"+nome+"%");
+		statement.setLong(2, userLogado);
+		// seta os valores dentro da variavel que foi pego no banco.
+		ResultSet resultado = statement.executeQuery();
+
+		while (resultado.next()) { // Percorre o objeto
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setSenha(resultado.getString("senha"));
+			modelLogin.setPerfil(resultado.getString("perfil"));
+			modelLogin.setSexo(resultado.getString("sexo"));
+			
+			retorno.add(modelLogin);
+		}
+
+		return retorno;
+	}
 	
-	public ModelLogin consultaUsuarioList(String login) throws Exception {
+	
+	public ModelLogin consultarUsuarioLogado(String login) throws Exception {
 
 		ModelLogin modelLogin = new ModelLogin();
 
-		String sql = "select * from model_login where upper(login) = upper('"+ login+"')";
+		String sql = "select * from model_login where upper(login) = upper('"+login+"')";
+		PreparedStatement statement = connection.prepareStatement(sql);
+		// seta os valores dentro da variavel que foi pego no banco.
+		ResultSet resultado = statement.executeQuery();
+
+		while (resultado.next()) { // Percorre o objeto
+			modelLogin.setId(resultado.getLong("id"));
+			modelLogin.setNome(resultado.getString("nome"));
+			modelLogin.setEmail(resultado.getString("email"));
+			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setSenha(resultado.getString("senha"));
+			modelLogin.setUseradmin(resultado.getBoolean("useradmin"));
+			modelLogin.setPerfil(resultado.getString("perfil"));
+			modelLogin.setSexo(resultado.getString("sexo"));
+			modelLogin.setFotouser(resultado.getString("fotouser"));
+		}
+
+		return modelLogin;
+	}
+	
+	
+	public ModelLogin consultarUsuario(String login) throws Exception {
+
+		ModelLogin modelLogin = new ModelLogin();
+
+		String sql = "select * from model_login where upper(login) = upper('"+login+"') and useradmin is false";
 
 		PreparedStatement statement = connection.prepareStatement(sql);
 		// seta os valores dentro da variavel que foi pego no banco.
@@ -110,20 +178,25 @@ public class DAOUsuarioRepository {
 			modelLogin.setEmail(resultado.getString("email"));
 			modelLogin.setLogin(resultado.getString("login"));
 			modelLogin.setSenha(resultado.getString("senha"));
+			modelLogin.setPerfil(resultado.getString("perfil"));
+			modelLogin.setSexo(resultado.getString("sexo"));
+			modelLogin.setFotouser(resultado.getString("fotouser"));
+
 		}
 
 		return modelLogin;
 	}
 	
 	
-	public ModelLogin consultaUsuarioID(String id) throws Exception  {
+	public ModelLogin consultaUsuarioID(String id, Long userLogado) throws Exception  {
 		
 		ModelLogin modelLogin = new ModelLogin();
 		
-		String sql = "select * from model_login where id = ? ";
+		String sql = "select * from model_login where id = ? and useradmin is false and usuario_id = ?";
 		
 		PreparedStatement statement = connection.prepareStatement(sql);
 		statement.setLong(1, Long.parseLong(id));
+		statement.setLong(2, userLogado);
 		
 		ResultSet resultado = statement.executeQuery();
 		
@@ -134,16 +207,19 @@ public class DAOUsuarioRepository {
 			modelLogin.setEmail(resultado.getString("email"));
 			modelLogin.setSenha(resultado.getString("senha"));
 			modelLogin.setLogin(resultado.getString("login"));
+			modelLogin.setPerfil(resultado.getString("perfil"));
+			modelLogin.setSexo(resultado.getString("sexo"));
+			modelLogin.setFotouser(resultado.getString("fotouser"));
 		}
 		
 		return modelLogin;
 	}
 	
-
+	
 	public boolean validarLogin(ModelLogin login) throws Exception {
 		
 		/*O count ele retorna os registro dentro de uma determinada coluna
-		 * Aqui ele estÃ¡ fazendo uma verificaÃ§Ã£o vendo se contem um usuario dentro da coluna
+		 * Aqui ele está fazendo uma verificação vendo se contem um usuario dentro da coluna
 		 * O maior que zero e para retorna um TRUE ou FALSE inves do ID*/
 		String sql = "select count(1) > 0 as existe from model_login where upper(login) = upper('"+login.getLogin()+"');";
 
@@ -157,7 +233,7 @@ public class DAOUsuarioRepository {
 	
 	public void deletarUser(String idUser) throws Exception {
 		
-		String sql = "DELETE FROM model_login WHERE id = ?;";
+		String sql = "DELETE FROM model_login WHERE id = ? and useradmin is false";
 		
 		PreparedStatement deleteUser =  connection.prepareStatement(sql);
 		deleteUser.setLong(1, Long.parseLong(idUser ));
